+++ tests/testthat/test-create_single_dose_dataset.R
@@ -1,18 +1,27 @@
 test_that("create_single_dose_dataset works as expected for Q#/EVERY # cases", {
   input <- tibble::tribble(
     ~USUBJID, ~EXDOSFRQ, ~ASTDT, ~ASTDTM, ~AENDT, ~AENDTM,
-    "P01", "Q2D", ymd("2021-01-01"), ymd_hms("2021-01-01 10:30:00"), ymd("2021-01-07"), ymd_hms("2021-01-07 11:30:00"),
-    "P01", "Q3D", ymd("2021-01-08"), ymd_hms("2021-01-08 12:00:00"), ymd("2021-01-14"), ymd_hms("2021-01-14 14:00:00"),
-    "P01", "EVERY 2 WEEKS", ymd("2021-01-15"), ymd_hms("2021-01-15 09:57:00"), ymd("2021-01-29"), ymd_hms("2021-01-29 10:57:00")
+    "P01", "Q2D", ymd("2021-01-01"), ymd_hms("2021-01-01 10:30:00"),
+     ymd("2021-01-07"), ymd_hms("2021-01-07 11:30:00"),
+    "P01", "Q3D", ymd("2021-01-08"), ymd_hms("2021-01-08 12:00:00"),
+     ymd("2021-01-14"), ymd_hms("2021-01-14 14:00:00"),
+    "P01", "EVERY 2 WEEKS", ymd("2021-01-15"), ymd_hms("2021-01-15 09:57:00"),
+     ymd("2021-01-29"), ymd_hms("2021-01-29 10:57:00")
   )
   expected_output <- tibble::tribble(
     ~USUBJID, ~EXDOSFRQ, ~ASTDT, ~ASTDTM, ~AENDT, ~AENDTM,
-    "P01", "ONCE", lubridate::ymd("2021-01-01"), lubridate::ymd_hms("2021-01-01 10:30:00"), lubridate::ymd("2021-01-01"), lubridate::ymd_hms("2021-01-01 11:30:00"),
-    "P01", "ONCE", lubridate::ymd("2021-01-03"), lubridate::ymd_hms("2021-01-03 10:30:00"), lubridate::ymd("2021-01-03"), lubridate::ymd_hms("2021-01-03 11:30:00"),
-    "P01", "ONCE", lubridate::ymd("2021-01-08"), lubridate::ymd_hms("2021-01-08 10:30:00"), lubridate::ymd("2021-01-08"), lubridate::ymd_hms("2021-01-08 11:30:00"),
-    "P01", "ONCE", lubridate::ymd("2021-01-11"), lubridate::ymd_hms("2021-01-11 10:30:00"), lubridate::ymd("2021-01-11"), lubridate::ymd_hms("2021-01-11 11:30:00"),
-    "P01", "ONCE", lubridate::ymd("2021-01-15"), lubridate::ymd_hms("2021-01-15 10:30:00"), lubridate::ymd("2021-01-15"), lubridate::ymd_hms("2021-01-15 11:30:00"),
-    "P01", "ONCE", lubridate::ymd("2021-01-29"), lubridate::ymd_hms("2021-01-29 10:30:00"), lubridate::ymd("2021-01-29"), lubridate::ymd_hms("2021-01-29 11:30:00")
+    "P01", "ONCE", lubridate::ymd("2021-01-01"), lubridate::ymd_hms("2021-01-01 10:30:00"),
+     lubridate::ymd("2021-01-01"), lubridate::ymd_hms("2021-01-01 11:30:00"),
+    "P01", "ONCE", lubridate::ymd("2021-01-03"), lubridate::ymd_hms("2021-01-03 10:30:00"),
+     lubridate::ymd("2021-01-03"), lubridate::ymd_hms("2021-01-03 11:30:00"),
+    "P01", "ONCE", lubridate::ymd("2021-01-08"), lubridate::ymd_hms("2021-01-08 10:30:00"),
+     lubridate::ymd("2021-01-08"), lubridate::ymd_hms("2021-01-08 11:30:00"),
+    "P01", "ONCE", lubridate::ymd("2021-01-11"), lubridate::ymd_hms("2021-01-11 10:30:00"),
+     lubridate::ymd("2021-01-11"), lubridate::ymd_hms("2021-01-11 11:30:00"),
+    "P01", "ONCE", lubridate::ymd("2021-01-15"), lubridate::ymd_hms("2021-01-15 10:30:00"),
+     lubridate::ymd("2021-01-15"), lubridate::ymd_hms("2021-01-15 11:30:00"),
+    "P01", "ONCE", lubridate::ymd("2021-01-29"), lubridate::ymd_hms("2021-01-29 10:30:00"),
+     lubridate::ymd("2021-01-29"), lubridate::ymd_hms("2021-01-29 11:30:00")
   )
 
   expect_equal(create_single_dose_dataset(input), expected_output)
@@ -22,33 +31,63 @@ test_that("create_single_dose_dataset works as expected for Q#/EVERY # cases", {
 test_that("create_single_dose_dataset works as expected for # TIMES PER cases", {
   input <- tibble::tribble(
     ~USUBJID, ~EXDOSFRQ, ~ASTDT, ~ASTDTM, ~AENDT, ~AENDTM,
-    "P01", "2 TIMES PER YEAR", lubridate::ymd("2021-01-01"), lubridate::ymd_hms("2021-01-01 10:00:00"), lubridate::ymd("2021-07-01"), lubridate::ymd_hms("2021-07-01 10:00:00"),
-    "P02", "2 TIMES PER YEAR", lubridate::ymd("2021-01-01"), lubridate::ymd_hms("2021-01-01 10:30:00"), lubridate::ymd("2021-12-31"), lubridate::ymd_hms("2021-12-31 10:30:00"),
-    "P03", "4 TIMES PER MONTH", lubridate::ymd("2021-02-01"), lubridate::ymd_hms("2021-02-01 11:00:00"), lubridate::ymd("2021-03-01"), lubridate::ymd_hms("2021-03-01 11:00:00"),
-    "P04", "4 TIMES PER MONTH", lubridate::ymd("2021-01-01"), lubridate::ymd_hms("2021-01-01 11:30:00"), lubridate::ymd("2021-01-20"), lubridate::ymd_hms("2021-01-20 11:00:00"),
-    "P05", "5 TIMES PER WEEK", lubridate::ymd("2021-01-15"), lubridate::ymd_hms("2021-01-15 12:00:00"), lubridate::ymd("2021-01-17"), lubridate::ymd_hms("2021-01-17 12:00:00"),
-    "P06", "5 TIMES PER WEEK", lubridate::ymd("2021-01-15"), lubridate::ymd_hms("2021-01-01 12:30:00"), lubridate::ymd("2021-01-21"), lubridate::ymd_hms("2021-01-21 12:00:00"),
+    "P01", "2 TIMES PER YEAR",
+     lubridate::ymd("2021-01-01"), lubridate::ymd_hms("2021-01-01 10:00:00"),
+     lubridate::ymd("2021-07-01"), lubridate::ymd_hms("2021-07-01 10:00:00"),
+    "P02", "2 TIMES PER YEAR",
+     lubridate::ymd("2021-01-01"), lubridate::ymd_hms("2021-01-01 10:30:00"),
+     lubridate::ymd("2021-12-31"), lubridate::ymd_hms("2021-12-31 10:30:00"),
+    "P03", "4 TIMES PER MONTH",
+     lubridate::ymd("2021-02-01"), lubridate::ymd_hms("2021-02-01 11:00:00"),
+     lubridate::ymd("2021-03-01"), lubridate::ymd_hms("2021-03-01 11:00:00"),
+    "P04", "4 TIMES PER MONTH",
+     lubridate::ymd("2021-01-01"), lubridate::ymd_hms("2021-01-01 11:30:00"),
+     lubridate::ymd("2021-01-20"), lubridate::ymd_hms("2021-01-20 11:00:00"),
+    "P05", "5 TIMES PER WEEK",
+     lubridate::ymd("2021-01-15"), lubridate::ymd_hms("2021-01-15 12:00:00"),
+     lubridate::ymd("2021-01-17"), lubridate::ymd_hms("2021-01-17 12:00:00"),
+    "P06", "5 TIMES PER WEEK",
+     lubridate::ymd("2021-01-15"), lubridate::ymd_hms("2021-01-01 12:30:00"),
+     lubridate::ymd("2021-01-21"), lubridate::ymd_hms("2021-01-21 12:00:00"),
   )
   expected_output <- tibble::tribble(
     ~USUBJID, ~EXDOSFRQ, ~ASTDT, ~ASTDTM, ~AENDT, ~AENDTM,
-    "P01", "ONCE", lubridate::ymd("2021-01-01"), lubridate::ymd_hms("2021-01-01 10:00:00"), lubridate::ymd("2021-01-01"), lubridate::ymd_hms("2021-01-01 10:00:00"),
-    "P02", "ONCE", lubridate::ymd("2021-01-01"), lubridate::ymd_hms("2021-01-01 10:30:00"), lubridate::ymd("2021-01-01"), lubridate::ymd_hms("2021-01-01 10:30:00"),
-    "P02", "ONCE", lubridate::ymd("2021-07-02"), lubridate::ymd_hms("2021-07-02 10:30:00"), lubridate::ymd("2021-07-02"), lubridate::ymd_hms("2021-01-01 10:30:00"),
-    "P03", "ONCE", lubridate::ymd("2021-02-01"), lubridate::ymd_hms("2021-02-01 11:00:00"), lubridate::ymd("2021-02-01"), lubridate::ymd_hms("2021-02-01 11:00:00"),
-    "P03", "ONCE", lubridate::ymd("2021-02-08"), lubridate::ymd_hms("2021-02-08 11:00:00"), lubridate::ymd("2021-02-08"), lubridate::ymd_hms("2021-02-08 11:00:00"),
-    "P03", "ONCE", lubridate::ymd("2021-02-16"), lubridate::ymd_hms("2021-02-16 11:00:00"), lubridate::ymd("2021-02-16"), lubridate::ymd_hms("2021-02-16 11:00:00"),
-    "P03", "ONCE", lubridate::ymd("2021-02-23"), lubridate::ymd_hms("2021-02-23 11:00:00"), lubridate::ymd("2021-02-23"), lubridate::ymd_hms("2021-02-23 11:00:00"),
-    "P04", "ONCE", lubridate::ymd("2021-01-01"), lubridate::ymd_hms("2021-01-01 11:30:00"), lubridate::ymd("2021-01-01"), lubridate::ymd_hms("2021-01-01 11:30:00"),
-    "P04", "ONCE", lubridate::ymd("2021-01-08"), lubridate::ymd_hms("2021-01-08 11:30:00"), lubridate::ymd("2021-01-08"), lubridate::ymd_hms("2021-01-08 11:30:00"),
-    "P04", "ONCE", lubridate::ymd("2021-01-16"), lubridate::ymd_hms("2021-01-16 11:30:00"), lubridate::ymd("2021-01-16"), lubridate::ymd_hms("2021-01-16 11:30:00"),
-    "P05", "ONCE", lubridate::ymd("2021-01-15"), lubridate::ymd_hms("2021-01-15 12:00:00"), lubridate::ymd("2021-01-15"), lubridate::ymd_hms("2021-01-15 12:00:00"),
-    "P05", "ONCE", lubridate::ymd("2021-01-16"), lubridate::ymd_hms("2021-01-16 12:00:00"), lubridate::ymd("2021-01-16"), lubridate::ymd_hms("2021-01-16 12:00:00"),
-    "P05", "ONCE", lubridate::ymd("2021-01-17"), lubridate::ymd_hms("2021-01-17 12:00:00"), lubridate::ymd("2021-01-17"), lubridate::ymd_hms("2021-01-17 12:00:00"),
-    "P06", "ONCE", lubridate::ymd("2021-01-15"), lubridate::ymd_hms("2021-01-15 12:30:00"), lubridate::ymd("2021-01-15"), lubridate::ymd_hms("2021-01-15 12:30:00"),
-    "P06", "ONCE", lubridate::ymd("2021-01-16"), lubridate::ymd_hms("2021-01-16 12:30:00"), lubridate::ymd("2021-01-16"), lubridate::ymd_hms("2021-01-16 12:30:00"),
-    "P06", "ONCE", lubridate::ymd("2021-01-17"), lubridate::ymd_hms("2021-01-17 12:30:00"), lubridate::ymd("2021-01-17"), lubridate::ymd_hms("2021-01-17 12:30:00"),
-    "P06", "ONCE", lubridate::ymd("2021-01-19"), lubridate::ymd_hms("2021-01-19 12:30:00"), lubridate::ymd("2021-01-19"), lubridate::ymd_hms("2021-01-19 12:30:00"),
-    "P06", "ONCE", lubridate::ymd("2021-01-20"), lubridate::ymd_hms("2021-01-20 12:30:00"), lubridate::ymd("2021-01-20"), lubridate::ymd_hms("2021-01-20 12:30:00")
+    "P01", "ONCE", lubridate::ymd("2021-01-01"), lubridate::ymd_hms("2021-01-01 10:00:00"),
+     lubridate::ymd("2021-01-01"), lubridate::ymd_hms("2021-01-01 10:00:00"),
+    "P02", "ONCE", lubridate::ymd("2021-01-01"), lubridate::ymd_hms("2021-01-01 10:30:00"),
+     lubridate::ymd("2021-01-01"), lubridate::ymd_hms("2021-01-01 10:30:00"),
+    "P02", "ONCE", lubridate::ymd("2021-07-02"), lubridate::ymd_hms("2021-07-02 10:30:00"),
+     lubridate::ymd("2021-07-02"), lubridate::ymd_hms("2021-01-01 10:30:00"),
+    "P03", "ONCE", lubridate::ymd("2021-02-01"), lubridate::ymd_hms("2021-02-01 11:00:00"),
+     lubridate::ymd("2021-02-01"), lubridate::ymd_hms("2021-02-01 11:00:00"),
+    "P03", "ONCE", lubridate::ymd("2021-02-08"), lubridate::ymd_hms("2021-02-08 11:00:00"),
+     lubridate::ymd("2021-02-08"), lubridate::ymd_hms("2021-02-08 11:00:00"),
+    "P03", "ONCE", lubridate::ymd("2021-02-16"), lubridate::ymd_hms("2021-02-16 11:00:00"),
+     lubridate::ymd("2021-02-16"), lubridate::ymd_hms("2021-02-16 11:00:00"),
+    "P03", "ONCE", lubridate::ymd("2021-02-23"), lubridate::ymd_hms("2021-02-23 11:00:00"),
+     lubridate::ymd("2021-02-23"), lubridate::ymd_hms("2021-02-23 11:00:00"),
+    "P04", "ONCE", lubridate::ymd("2021-01-01"), lubridate::ymd_hms("2021-01-01 11:30:00"),
+     lubridate::ymd("2021-01-01"), lubridate::ymd_hms("2021-01-01 11:30:00"),
+    "P04", "ONCE", lubridate::ymd("2021-01-08"), lubridate::ymd_hms("2021-01-08 11:30:00"),
+     lubridate::ymd("2021-01-08"), lubridate::ymd_hms("2021-01-08 11:30:00"),
+    "P04", "ONCE", lubridate::ymd("2021-01-16"), lubridate::ymd_hms("2021-01-16 11:30:00"),
+     lubridate::ymd("2021-01-16"), lubridate::ymd_hms("2021-01-16 11:30:00"),
+    "P05", "ONCE", lubridate::ymd("2021-01-15"), lubridate::ymd_hms("2021-01-15 12:00:00"),
+     lubridate::ymd("2021-01-15"), lubridate::ymd_hms("2021-01-15 12:00:00"),
+    "P05", "ONCE", lubridate::ymd("2021-01-16"), lubridate::ymd_hms("2021-01-16 12:00:00"),
+     lubridate::ymd("2021-01-16"), lubridate::ymd_hms("2021-01-16 12:00:00"),
+    "P05", "ONCE", lubridate::ymd("2021-01-17"), lubridate::ymd_hms("2021-01-17 12:00:00"),
+     lubridate::ymd("2021-01-17"), lubridate::ymd_hms("2021-01-17 12:00:00"),
+    "P06", "ONCE", lubridate::ymd("2021-01-15"), lubridate::ymd_hms("2021-01-15 12:30:00"),
+     lubridate::ymd("2021-01-15"), lubridate::ymd_hms("2021-01-15 12:30:00"),
+    "P06", "ONCE", lubridate::ymd("2021-01-16"), lubridate::ymd_hms("2021-01-16 12:30:00"),
+     lubridate::ymd("2021-01-16"), lubridate::ymd_hms("2021-01-16 12:30:00"),
+    "P06", "ONCE", lubridate::ymd("2021-01-17"), lubridate::ymd_hms("2021-01-17 12:30:00"),
+     lubridate::ymd("2021-01-17"), lubridate::ymd_hms("2021-01-17 12:30:00"),
+    "P06", "ONCE", lubridate::ymd("2021-01-19"), lubridate::ymd_hms("2021-01-19 12:30:00"),
+     lubridate::ymd("2021-01-19"), lubridate::ymd_hms("2021-01-19 12:30:00"),
+    "P06", "ONCE", lubridate::ymd("2021-01-20"), lubridate::ymd_hms("2021-01-20 12:30:00"),
+     lubridate::ymd("2021-01-20"), lubridate::ymd_hms("2021-01-20 12:30:00")
   )
 
   expect_equal(create_single_dose_dataset(input), expected_output)
@@ -57,8 +96,10 @@ test_that("create_single_dose_dataset works as expected for # TIMES PER cases", 
 test_that("create_single_dose_dataset works for different treatments", {
   input <- tibble::tribble(
     ~USUBJID, ~EXDOSFRQ, ~ASTDT, ~ASTDTM, ~AENDT, ~AENDT, ~EXTRT,
-    "P01", "Q2D", lubridate::ymd("2021-01-01"), lubridate::ymd_hms("2021-01-01 09:00:00"), lubridate::ymd("2021-01-03"), lubridate::ymd_hms("2021-01-03 09:00:00"), "XANOMELINE",
-    "P01", "QOD", lubridate::ymd("2021-01-01"), lubridate::ymd_hms("2021-01-01 09:15:00"), lubridate::ymd("2021-01-05"), lubridate::ymd_hms("2021-01-05 09:15:00"), "PLACEBO"
+    "P01", "Q2D", lubridate::ymd("2021-01-01"), lubridate::ymd_hms("2021-01-01 09:00:00"),
+     lubridate::ymd("2021-01-03"), lubridate::ymd_hms("2021-01-03 09:00:00"), "XANOMELINE",
+    "P01", "QOD", lubridate::ymd("2021-01-01"), lubridate::ymd_hms("2021-01-01 09:15:00"),
+     lubridate::ymd("2021-01-05"), lubridate::ymd_hms("2021-01-05 09:15:00"), "PLACEBO"
   )
   expected_output <- tibble::tribble(
     ~USUBJID, ~EXDOSFRQ, ~ASTDT, ~ASTDTM, ~AENDT, ~AENDTM, ~EXTRT,
@@ -75,7 +116,7 @@ test_that("create_single_dose_dataset works for different treatments", {
   )
 
   expect_equal(create_single_dose_dataset(input,
-                                          keep_source_vars = vars(USUBJID, EXDOSFRQ, ASTDT, ASTDTM, AENDT, AENDTM)
+                keep_source_vars = vars(USUBJID, EXDOSFRQ, ASTDT, ASTDTM, AENDT, AENDTM)
                                          ),
                expected_output)
 })
@@ -123,9 +164,12 @@ test_that("custom lookup works", {
 test_that("Warning is returned when values in EXDOSFRQ does not appear in lookup table", {
   input <- tibble::tribble(
     ~USUBJID, ~EXDOSFRQ, ~ASTDT, ~ASTDTM, ~AENDT, ~AENDTM,
-    "P01", "1", lubridate::ymd("2021-01-01"), lubridate::ymd_hms("2021-01-01T09:00:00"), lubridate::ymd("2021-01-03"), lubridate::ymd_hms("2021-01-03T09:00:00"),
-    "P01", "1", lubridate::ymd("2021-01-08"), lubridate::ymd_hms("2021-01-08T09:00:00"), lubridate::ymd("2021-01-12"), lubridate::ymd_hms("2021-01-12T09:00:00"),
-    "P01", "1", lubridate::ymd("2021-01-15"), lubridate::ymd_hms("2021-01-15T09:00:00"), lubridate::ymd("2021-01-29"), lubridate::ymd_hms("2021-01-29T09:00:00")
+    "P01", "1", lubridate::ymd("2021-01-01"), lubridate::ymd_hms("2021-01-01T09:00:00"),
+     lubridate::ymd("2021-01-03"), lubridate::ymd_hms("2021-01-03T09:00:00"),
+    "P01", "1", lubridate::ymd("2021-01-08"), lubridate::ymd_hms("2021-01-08T09:00:00"),
+     lubridate::ymd("2021-01-12"), lubridate::ymd_hms("2021-01-12T09:00:00"),
+    "P01", "1", lubridate::ymd("2021-01-15"), lubridate::ymd_hms("2021-01-15T09:00:00"),
+     lubridate::ymd("2021-01-29"), lubridate::ymd_hms("2021-01-29T09:00:00")
   )
   expect_error(
     create_single_dose_dataset(input)
@@ -134,10 +178,12 @@ test_that("Warning is returned when values in EXDOSFRQ does not appear in lookup
 
 test_that("Error is returned when a date variable contains NA values", {
   input <- tibble::tribble(
-    ~USUBJID, ~EXDOSFRQ, ~ASTDT, ~ASTDTM, ~AENDT, ~AENDTM
+    ~USUBJID, ~EXDOSFRQ, ~ASTDT, ~ASTDTM, ~AENDT, ~AENDTM,
     "P01", "Q2D", ymd("2021-01-01"), lubridate::ymd_hms("2021-01-01T09:00:00"), NA, NA,
-    "P01", "Q3D", ymd("2021-01-08"), lubridate::ymd_hms("2021-01-08T09:00:00"), ymd("2021-01-15"),lubridate::ymd_hms("2021-01-15T09:00:00"),
-    "P01", "EVERY 2 WEEKS", ymd("2021-01-15"), lubridate::ymd_hms("2021-01-15T09:00:00"), ymd("2021-01-29"), lubridate::ymd_hms("2021-01-29T09:00:00")
+    "P01", "Q3D", ymd("2021-01-08"), lubridate::ymd_hms("2021-01-08T09:00:00"),
+     ymd("2021-01-15"), lubridate::ymd_hms("2021-01-15T09:00:00"),
+    "P01", "EVERY 2 WEEKS", ymd("2021-01-15"), lubridate::ymd_hms("2021-01-15T09:00:00"),
+     ymd("2021-01-29"), lubridate::ymd_hms("2021-01-29T09:00:00")
   )
   expect_error(
     create_single_dose_dataset(input),